requirements.txt
Django==2.0.9
Whoosh==2.7.4
django-auth-ldap==1.5.0
django-form-utils==1.0.3
django-haystack==2.8.1
django-reversion==2.0.13
django-tastypie==0.14.1
psycopg2-binary
elasticsearch>=2.0.0,<3.0.0 # rq.filter: >=2.0.0,<3.0.0
pytz
requests==2.20.0
simplejson==3.14.0
six==1.11.0
djangorestframework==3.8.2
django-filter==1.1.0
markdown==2.6.11
django_compressor==2.2
python-memcached==1.59
django-debug-toolbar==1.9.1
django-reversion-compare==0.8.4
diff-match-patch==20121119
django-enumfield==1.2.1
django-bootstrap3==10.0.1
python-ldap==3.0.0
python-dateutil==2.7.2
python-mimeparse==1.6.0
whitenoise==3.3.1
django-health-check==3.5.1
gunicorn
ddtrace
